+++ CHANGELOG.md
@@ -1,5 +1,11 @@
 # Change Log
 
+## [0.5.0](https://github.com/wbrandenburger/PSPocs/tree/0.5.0) (2019-09-22)
+
+**Implemented enhancements:**
+
+- Added parameter `Name`, `VirtualEnv` and `Identifier` to function `New-PocsLibrary` for generated adjusted libraries.
+
 ## [0.4.7](https://github.com/wbrandenburger/PSPocs/tree/0.4.7) (2019-09-22)
 
 **Implemented enhancements:**